LayoutTests/fast/frames/adopt-iframe-into-itself-expected.txt

ALERT: Error: HIERARCHY_REQUEST_ERR: DOM Exception 3
